refactor(sonar): close 23 of 27 new-code smells
CRITICAL S3776 - StockQueryService.fetchPage had cognitive complexity 17.
  Extract the parallel fetch into fetchSummariesInParallel() and split
  the inner steps (submitSummaryFutures, safeFetchSummary,
  collectSummaryFutures). fetchPage is now an orchestrator of ~3 lines.

MAJOR S6878 - ImageProxyService.ProxiedImage.equals now uses the Java 21
  record pattern (instanceof Success(byte[] otherContent, ...)) instead
  of named binding + field access.

MAJOR S5778 - PortfolioHoldingsBuilderTest.assertThatThrownBy lambda
  hoisted the List.of(s) construction out so the lambda only contains
  the actual call under test.

MINOR S1905 - NewsAggregatorService dropped the (long) cast in
  'new NewsQueryResult(..., (long) matched.size(), ...)'; int->long
  widening is implicit.

MINOR S1128 - PriceHistoryToolTest dropped the unused AssetType import.

MINOR S5853 (x ~14) - consecutive AssertJ assertions on the same subject
  in tool tests (Alarm/Watchlist/Price*/Portfolio/News) merged into
  fluent chains for readability and to satisfy the rule.

Remaining (acceptable):
  - 3x S5976 (suggest @ParameterizedTest) - kept as individual tests
    because each test's @DisplayName captures a distinct scenario.
  - S1135 INFO on BankCurrencyController TODO - intentional marker.

Verified: 271 unit + 2 IT, 0 failures, 0 regressions.

+++ backend/finance-portal-backend/src/main/java/com/finance/portal/news/application/service/NewsAggregatorService.java
@@ -191,7 +191,8 @@ public class NewsAggregatorService {
                 .toList();
 
         List<NewsArticle> out = translatePage(matched, lang);
-        return new NewsQueryResult(out, 1, size, (long) matched.size(), 1, List.of(), Map.of());
+        // int → long widening implicit; explicit cast gereksiz (S1905).
+        return new NewsQueryResult(out, 1, size, matched.size(), 1, List.of(), Map.of());
     }
 
     private static int scoreFor(NewsArticle a, Pattern kwPattern, Set<NewsCategory> cats) {
